fix: split app smoke task into ci_app_task.py to avoid fastapi import in hello pod
When the hello task pod starts, it imports ci_smoke_task.py (uploaded via
fast registration). The module-level _make_fastapi_app() call triggered
`import fastapi`, which is not in the hello environment (_smoke_env uses
the default image with no extra packages). This caused a
ModuleNotFoundError and the hello run to fail.

Fix: extract all FastAPI-dependent code into ci_app_task.py. ci_smoke_task.py
now imports only flyte (no flyte.app.extras). ci_app_task.py is only loaded
by the app-tester pod whose image includes fastapi, uvicorn, and httpx.

Also removed fastapi from _imgbuild_env and _imgcache_env pip_packages since
neither task actually uses fastapi — they only test requests.
